Added sample Demand and Cost Matrix
Added sample Demand and Cost Matrix files for Material Supply Problem Flask app. These files can be used as input to the app.py program for solving the transportation problem. The Demand file contains a list of demand nodes with their respective demand quantities and the Cost Matrix file contains the costs of transporting one unit of material from each supply node to each demand node. These files are added to the repo for ease of use for the users.